Extract swap settings management to seperate view model
