+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/swapxxx/settings/SwapSettingsFragment.kt
@@ -32,10 +32,12 @@ class SwapSettingsFragment : BaseComposeFragment() {
 
 @Composable
 private fun SwapProviderSettingsScreen(navController: NavController) {
-    val viewModel = viewModel<SwapViewModel>(
+    val swapViewModel = viewModel<SwapViewModel>(
         viewModelStoreOwner = navController.previousBackStackEntry!!,
         factory = SwapViewModel.Factory()
     )
+
+    val viewModel = viewModel<SwapSettingsViewModel>(factory = SwapSettingsViewModel.Factory())
 
     Scaffold(
         topBar = {
@@ -64,6 +66,7 @@ private fun SwapProviderSettingsScreen(navController: NavController) {
                     enabled = viewModel.saveSettingsEnabled,
                     onClick = {
                         viewModel.saveSettings()
+                        swapViewModel.onUpdateSettings(viewModel.getSettings())
                         navController.popBackStack()
                     }
                 )
@@ -78,7 +81,7 @@ private fun SwapProviderSettingsScreen(navController: NavController) {
                 VSpacer(height = 12.dp)
             }
 
-            viewModel.uiState.quote?.swapQuote?.let { swapQuote ->
+            swapViewModel.uiState.quote?.swapQuote?.let { swapQuote ->
                 items(swapQuote.settings) { setting ->
                     val settingId = setting.id
